requirements.txt
django==2
pillow
django-crispy-forms
paywix
requests
django-simple-history
django-extensions
img2pdf
text2pdf
django-widget-tweaks
psycopg2
PyPDF2
